+++ openair3/NAS/COMMON/EMM/MSG/AuthenticationFailure.c
@@ -35,7 +35,7 @@ int decode_authentication_failure(authentication_failure_msg *authentication_fai
   int decoded_result = 0;
 
   // Check if we got a NULL pointer and if buffer length is >= minimum length expected for the message.
-  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, AUTHENTICATION_FAILURE_MINIMUM_LENGTH, len);
+  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, AUTHENTICATION_FAILURE_MINIMUM_LENGTH, (int)len);
 
   /* Decoding mandatory fields */
   if ((decoded_result = decode_emm_cause(&authentication_failure->emmcause, 0, buffer + decoded, len - decoded)) < 0)
@@ -79,7 +79,7 @@ int encode_authentication_failure(authentication_failure_msg *authentication_fai
   int encode_result = 0;
 
   /* Checking IEI and pointer */
-  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, AUTHENTICATION_FAILURE_MINIMUM_LENGTH, len);
+  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, AUTHENTICATION_FAILURE_MINIMUM_LENGTH, (int)len);
 
   if ((encode_result = encode_emm_cause(&authentication_failure->emmcause, 0,
                                         buffer + encoded, len - encoded)) < 0)        //Return in case of error
