+++ openair3/NAS/COMMON/EMM/MSG/AuthenticationReject.c
@@ -33,7 +33,7 @@ int decode_authentication_reject(authentication_reject_msg *authentication_rejec
 {
   uint32_t decoded = 0;
   // Check if we got a NULL pointer and if buffer length is >= minimum length expected for the message.
-  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, AUTHENTICATION_REJECT_MINIMUM_LENGTH, len);
+  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, AUTHENTICATION_REJECT_MINIMUM_LENGTH, (int)len);
 
   /* Decoding mandatory fields */
   return decoded;
@@ -43,7 +43,7 @@ int encode_authentication_reject(authentication_reject_msg *authentication_rejec
 {
   int encoded = 0;
   /* Checking IEI and pointer */
-  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, AUTHENTICATION_REJECT_MINIMUM_LENGTH, len);
+  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, AUTHENTICATION_REJECT_MINIMUM_LENGTH, (int)len);
 
   return encoded;
 }
