+++ openair3/NAS/COMMON/EMM/MSG/EmmStatus.c
@@ -35,7 +35,7 @@ int decode_emm_status(emm_status_msg *emm_status, uint8_t *buffer, uint32_t len)
   int decoded_result = 0;
 
   // Check if we got a NULL pointer and if buffer length is >= minimum length expected for the message.
-  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, EMM_STATUS_MINIMUM_LENGTH, len);
+  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, EMM_STATUS_MINIMUM_LENGTH, (int)len);
 
   /* Decoding mandatory fields */
   if ((decoded_result = decode_emm_cause(&emm_status->emmcause, 0, buffer + decoded, len - decoded)) < 0)
@@ -52,7 +52,7 @@ int encode_emm_status(emm_status_msg *emm_status, uint8_t *buffer, uint32_t len)
   int encode_result = 0;
 
   /* Checking IEI and pointer */
-  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, EMM_STATUS_MINIMUM_LENGTH, len);
+  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, EMM_STATUS_MINIMUM_LENGTH, (int)len);
 
   if ((encode_result = encode_emm_cause(&emm_status->emmcause, 0, buffer +
                                         encoded, len - encoded)) < 0)        //Return in case of error
