+++ openair3/NAS/COMMON/EMM/MSG/SecurityModeReject.c
@@ -35,7 +35,7 @@ int decode_security_mode_reject(security_mode_reject_msg *security_mode_reject, 
   int decoded_result = 0;
 
   // Check if we got a NULL pointer and if buffer length is >= minimum length expected for the message.
-  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, SECURITY_MODE_REJECT_MINIMUM_LENGTH, len);
+  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, SECURITY_MODE_REJECT_MINIMUM_LENGTH, (int)len);
 
   /* Decoding mandatory fields */
   if ((decoded_result = decode_emm_cause(&security_mode_reject->emmcause, 0, buffer + decoded, len - decoded)) < 0)
@@ -52,7 +52,7 @@ int encode_security_mode_reject(security_mode_reject_msg *security_mode_reject, 
   int encode_result = 0;
 
   /* Checking IEI and pointer */
-  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, SECURITY_MODE_REJECT_MINIMUM_LENGTH, len);
+  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, SECURITY_MODE_REJECT_MINIMUM_LENGTH, (int)len);
 
   if ((encode_result = encode_emm_cause(&security_mode_reject->emmcause, 0,
                                         buffer + encoded, len - encoded)) < 0)        //Return in case of error
