+++ openair3/NAS/COMMON/ESM/MSG/ActivateDefaultEpsBearerContextReject.c
@@ -35,7 +35,7 @@ int decode_activate_default_eps_bearer_context_reject(activate_default_eps_beare
   int decoded_result = 0;
 
   // Check if we got a NULL pointer and if buffer length is >= minimum length expected for the message.
-  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, ACTIVATE_DEFAULT_EPS_BEARER_CONTEXT_REJECT_MINIMUM_LENGTH, len);
+  CHECK_PDU_POINTER_AND_LENGTH_DECODER(buffer, ACTIVATE_DEFAULT_EPS_BEARER_CONTEXT_REJECT_MINIMUM_LENGTH, (int)len);
 
   /* Decoding mandatory fields */
   if ((decoded_result = decode_esm_cause(&activate_default_eps_bearer_context_reject->esmcause, 0, buffer + decoded, len - decoded)) < 0)
@@ -79,7 +79,7 @@ int encode_activate_default_eps_bearer_context_reject(activate_default_eps_beare
   int encode_result = 0;
 
   /* Checking IEI and pointer */
-  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, ACTIVATE_DEFAULT_EPS_BEARER_CONTEXT_REJECT_MINIMUM_LENGTH, len);
+  CHECK_PDU_POINTER_AND_LENGTH_ENCODER(buffer, ACTIVATE_DEFAULT_EPS_BEARER_CONTEXT_REJECT_MINIMUM_LENGTH, (int)len);
 
   if ((encode_result =
          encode_esm_cause(&activate_default_eps_bearer_context_reject->esmcause,
